+++ Runtime/Data/Dialogue.cs
@@ -9,7 +9,7 @@ namespace OneM.DialogueSystem
     public sealed class Dialogue : ScriptableObject
     {
         [field: SerializeField, Tooltip("All Actors present in this dialogue.")]
-        public DialogueActor[] Actors { get; private set; }
+        public Actor Actor { get; private set; }
 
         [field: SerializeField, Tooltip("The lines used on this dialogue.")]
         public DialogueLine[] Lines { get; private set; }