+++ Runtime/UI/DialogueUIBoard.cs
@@ -2,27 +2,25 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.Localization.Components;
 using ActionCode.AwaitableSystem;
-using ActionCode.SerializedDictionaries;
 
 namespace OneM.DialogueSystem
 {
     [DisallowMultipleComponent]
     public sealed class DialogueUIBoard : MonoBehaviour
     {
+        [SerializeField] private DialogueUIActor actor;
         [SerializeField] private AudioSource audioSource;
+        [SerializeField] private GameObject marker;
+
+        [Header("LINES")]
         [SerializeField] private TMP_Text textLine;
         [SerializeField] private LocalizeStringEvent localizedLine;
-        [SerializeField] private GameObject completionLineMarker;
 
         [Header("TIMERS")]
         [SerializeField, Tooltip("The initial time to wait (in seconds) before start to writing.")]
         private float initialWaitingTime = 1f;
         [SerializeField, Tooltip("The time (in seconds) to wait between each letter.")]
         private float typeWriteTime = 0.02f;
-
-        [Space]
-        [SerializeField, Tooltip("The actors instances.")]
-        private SerializedDictionary<ActorPosition, DialogueUIActor> actors;
 
         public bool IsTypeWriting { get; private set; }
         public bool IsNextLineAvailable { get; private set; }
@@ -31,8 +29,8 @@ namespace OneM.DialogueSystem
 
         public async Awaitable PlayAsync(Dialogue dialogue)
         {
-            SetCompletionLineMarkerEnable(false);
-            LoadActors(dialogue.Actors);
+            actor.Load(dialogue.Actor);
+            SetMarkerEnable(false);
 
             gameObject.SetActive(true);
             await AwaitableUtility.WaitForSecondsRealtimeAsync(initialWaitingTime);
@@ -40,11 +38,9 @@ namespace OneM.DialogueSystem
             foreach (var line in dialogue.Lines)
             {
                 IsNextLineAvailable = false;
-
-                SetActorNameEnable(line.Position);
                 localizedLine.StringReference = line.LocalizedLine;
 
-                await WaitUntilLocalizedLineIsLoadedAsync();
+                await WaitUntilLocalizedLineIsFullyLoadedAsync();
                 await PlayTypeWriteLineAnimationAsync();
                 await WaitUntilNextLineIsAvailableAsync();
             }
@@ -69,13 +65,12 @@ namespace OneM.DialogueSystem
 
         public void Disable()
         {
+            actor.Dispose();
             gameObject.SetActive(false);
-            SetCompletionLineMarkerEnable(false);
+            SetMarkerEnable(false);
 
             textLine.text = string.Empty;
             localizedLine.StringReference = null;
-
-            DisposeActors();
         }
 
         private bool CanAdvance()
@@ -84,7 +79,7 @@ namespace OneM.DialogueSystem
             return framesSinceLastAdvance > 10;
         }
 
-        private async Awaitable WaitUntilLocalizedLineIsLoadedAsync() =>
+        private async Awaitable WaitUntilLocalizedLineIsFullyLoadedAsync() =>
             await localizedLine.StringReference.GetLocalizedStringAsync().Task;
 
         private async Awaitable PlayTypeWriteLineAnimationAsync()
@@ -94,7 +89,7 @@ namespace OneM.DialogueSystem
 
             audioSource.Play();
             IsTypeWriting = true;
-            SetCompletionLineMarkerEnable(false);
+            SetMarkerEnable(false);
 
             while (textLine.maxVisibleCharacters < textLength)
             {
@@ -104,41 +99,13 @@ namespace OneM.DialogueSystem
 
             audioSource.Stop();
             IsTypeWriting = false;
-            SetCompletionLineMarkerEnable(true);
+            SetMarkerEnable(true);
         }
 
         private async Awaitable WaitUntilNextLineIsAvailableAsync() =>
             await AwaitableUtility.WaitUntilAsync(() => IsNextLineAvailable);
 
         private void CompleteTypeWrite() => textLine.maxVisibleCharacters = textLine.text.Length;
-
-        private void LoadActors(DialogueActor[] dialogueActors)
-        {
-            foreach (var dialogue in dialogueActors)
-            {
-                var actor = actors[dialogue.Position];
-                actor.Load(dialogue.Actor);
-            }
-        }
-
-        private void SetActorNameEnable(ActorPosition position)
-        {
-            foreach (var actor in actors.Values)
-            {
-                actor.SetNameActive(false);
-            }
-            actors[position].SetNameActive(true);
-        }
-
-        private void SetCompletionLineMarkerEnable(bool isEnable) =>
-            completionLineMarker.SetActive(isEnable);
-
-        private void DisposeActors()
-        {
-            foreach (var actor in actors.Values)
-            {
-                actor.Dispose();
-            }
-        }
+        private void SetMarkerEnable(bool isEnable) => marker.SetActive(isEnable);
     }
 }